Add ping functionality to connections package
This commit adds a new file `ping.go` to the `connections` package, which contains the `Pinger` function. The `Pinger` function implements a ping functionality that sends a "ping" message to a writer at regular intervals. It also includes a timer that can be reset to change the interval duration.

Additionally, this commit adds two new test files `ping_example_test.go` and `ping_test.go` that test the `Pinger` function in different scenarios.